9.28.18
Written for Proving Ground server,

currently with Shine, and no Node concept.

Queue is small count <= 6, doesn't matter too much about for loops.

This process in no way shape or form is intended to be deceiving to server browser. The idea is to automate the process of waiting to join a team as spectator.

By default, the spectator must f4 then join team. with this, they join a queue and stay in spectator and get set to the team when a spot is open.

This adds functions for teamjoin to check when a spot is open based on the desired sh_teamcap size.

The only misleading portion is sh_teamcap command if less than the current slot limit count. This is due to the slot limit count being set at server launch, and not adjustable without server reboot.

If a spect f4s then joins team, that process is faster than my eyes see in server browser going from 24 to 23. This process is faster than human, so it is not misleading. It's a tool. It's helpful.

I had fun writing. I have to make sure the requirements fit the bill!